requirements.txt
PyPDF2>=3.0.1
python-docx>=0.8.11
prompt_toolkit>=3.0.47
requests>=2.31.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
openai>=0.28.1
anthropic>=0.2.9
